Added Authorisation and Authentication

+++ src/main/java/com/explorebnb/clone/airBnbApp/service/HotelServiceImpl.java
@@ -5,7 +5,9 @@ import com.explorebnb.clone.airBnbApp.dto.HotelInfoDto;
 import com.explorebnb.clone.airBnbApp.dto.RoomDto;
 import com.explorebnb.clone.airBnbApp.entity.Hotel;
 import com.explorebnb.clone.airBnbApp.entity.Room;
+import com.explorebnb.clone.airBnbApp.entity.User;
 import com.explorebnb.clone.airBnbApp.exception.ResourceNotFoundException;
+import com.explorebnb.clone.airBnbApp.exception.UnAuthorisedException;
 import com.explorebnb.clone.airBnbApp.repository.HotelRepository;
 import com.explorebnb.clone.airBnbApp.repository.InventoryRepository;
 import com.explorebnb.clone.airBnbApp.repository.RoomRepository;
@@ -13,6 +15,7 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -33,6 +36,8 @@ public class HotelServiceImpl implements HotelService {
     public HotelDto createNewHotel(HotelDto hotelDto) {
         log.info("Creating hotel with name:{}",hotelDto.getName());
         Hotel hotel=modelMapper.map(hotelDto,Hotel.class);
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
         hotel.setActive(false);
         hotel=hotelRepository.save(hotel);
         log.info("Created a new hotel with id:{}",hotelDto.getId());
@@ -45,6 +50,11 @@ public class HotelServiceImpl implements HotelService {
         Hotel hotel=hotelRepository.findById(id).orElseThrow(()->new
                 ResourceNotFoundException("Hotel with id:"+id+" not found"));
         log.info("Hotel with id:{}",id);
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner()))
+        {
+            throw new UnAuthorisedException("This user does not own this hotel with id:"+id);
+        }
         return modelMapper.map(hotel,HotelDto.class);
     }
 
@@ -53,6 +63,11 @@ public class HotelServiceImpl implements HotelService {
         log.info("Updating hotel details for id:{}",id);
         Hotel hotel=hotelRepository.findById(id).orElseThrow(()->
                 new ResourceNotFoundException("Hotel with id: "+id+" doesn't exist"));
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner()))
+        {
+            throw new UnAuthorisedException("This user does not own this hotel with id:"+id);
+        }
         modelMapper.map(hotelDto,hotel);
         hotel.setId(id);
         hotel=hotelRepository.save(hotel);
@@ -65,7 +80,11 @@ public class HotelServiceImpl implements HotelService {
         log.info("Deleting the hotel with id:{}",id);
         Hotel hotel=hotelRepository.findById(id).orElseThrow(()->
                 new ResourceNotFoundException("Hotel with id:{}"+id+" doesn't exist"));
-
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner()))
+        {
+            throw new UnAuthorisedException("This user does not own this hotel with id:"+id);
+        }
         for(Room room:hotel.getRooms()){
             inventoryService.deleteAllInventories(room);
             roomRepository.deleteById(room.getId());
@@ -81,9 +100,13 @@ public class HotelServiceImpl implements HotelService {
         log.info("Activating the Hotel with id:{}",hotelId);
         Hotel hotel=hotelRepository.findById(hotelId).orElseThrow(
                 ()->new ResourceNotFoundException("Hotel with id: "+hotelId+" doesn't exist."));
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner()))
+        {
+            throw new UnAuthorisedException("This user does not own this hotel with id:"+hotelId);
+        }
         hotel.setActive(true);
         //add the future inventories
-        hotelRepository.save(hotel);
         for(Room room:hotel.getRooms()){
             inventoryService.initializeRoomForAYear(room);
         }
